moved valid date function to utils
