+++ lac_validator/rules/rule_utils.py
@@ -73,3 +73,41 @@ def field_different_from_previous(dfs, field):
         err_list = merged_co["index"][err_mask].unique().tolist()
         err_list.sort()
         return {"Episodes": err_list}
+    
+def valid_date(dte):
+    try:
+        lst = dte.split("/")
+    except AttributeError:
+        return pd.NaT
+    # Preceding block checks for the scenario where the value passed in is nan/naT
+
+    # date should have three elements
+    if len(lst) != 3:
+        return pd.NaT
+
+    z_list = ["ZZ", "ZZ", "ZZZZ"]
+    # We set the date to the latest possible value to avoid false positives
+    offset_list = [
+        pd.DateOffset(months=1, days=-1),
+        pd.DateOffset(years=1, days=-1),
+        None,
+    ]
+    # that is, go to the next month/year and take the day before that
+    date_bits = []
+
+    for i, zeds, offset in zip(lst, z_list, offset_list):
+        if i == "ZZ":
+            i = "01"
+            offset_to_use = offset
+        elif i == "ZZZZ":
+            i = "2000"
+        date_bits.append(i)
+
+    as_datetime = pd.to_datetime(
+        "/".join(date_bits), format="%d/%m/%Y", errors="coerce"
+    )
+    try:
+        as_datetime += offset_to_use
+    except (NameError, TypeError):  # offset_to_use only defined if needed
+        pass
+    return as_datetime
